@@ -1,7 +1,8 @@
 {
   "name": "html2image",
   "version": "1.2.0",
-  "description": "convert html to image/pdf by phantomjs",
+  "description": "convert html to image/pdf by puppeteer",
+  "type": "module",
   "main": "screencapture.js",
   "scripts": {
     "test": "node test.js"
@@ -21,12 +22,12 @@
   "preferGlobal": true,
   "license": "MIT",
   "dependencies": {
-    "commander": "^2.8.1",
-    "mkdirp": "^0.5.1",
-    "ora": "^1.3.0",
-    "phantomjs-prebuilt": "^2.1.16"
+    "commander": "^8.3.0",
+    "mkdirp": "^1.0.4",
+    "ora": "^6.0.1",
+    "puppeteer": "^13.1.1"
   },
   "engines": {
-    "node": ">=6"
+    "node": ">=12.0.0"
   }
 }
