@@ -2,26 +2,26 @@
  * Created by huangxinghui on 2015/9/15.
  */
 
-const childProcess = require('child_process');
-const phantomjs = require('phantomjs-prebuilt');
-const ora = require('ora');
-const mkdirp = require('mkdirp');
+import ora from 'ora';
+import mkdirp from 'mkdirp';
+import pdfExporter from './lib/pdf.js';
+import imgExporter from './lib/img.js';
 
-function screenCapture(address, format, output) {
-    const spinner = ora(`Open address ${address}`).start();
+async function screenCapture(address, format, output) {
+  const spinner = ora(`Open address ${address}`).start();
 
-    mkdirp.sync(output);
+  mkdirp.sync(output);
 
-    var program = phantomjs.exec('./lib/screencapture.js', address, format, output);
-    program.stderr.pipe(process.stderr);
-    program.stdout.pipe(process.stdout);
-    program.on('exit', code => {
-        if (code === 0) {
-            spinner.succeed(`Convert address ${address} succeed`);
-        } else {
-            spinner.fail(`Load address ${address} fail`);
-        }
-    });
+  try {
+    if (format === 'pdf') {
+      await pdfExporter(address, output);
+    } else {
+      await imgExporter(address, output, format);
+    }
+    spinner.succeed(`Convert address ${address} succeed`);
+  } catch (e) {
+    spinner.fail(`Load address ${address} fail`);
+  }
 }
 
-module.exports = screenCapture;
+export default screenCapture;
