@@ -1,3 +1,3 @@
-let capture = require('./screencapture');
+let capture = require('./screencapture.js');
 
 capture('https://www.baidu.com/', 'pdf', '../test')
